# Optimize DstAggregator by caching remote allocations